build - Ability to specify the initial pom required more support.
* Ability to specify the intial pom broke existing builds which only
specified a -J argument and/or their -P argument did not explicitly
list the root project pom. These builds would fail as the foam pom
was not loaded. The build now addes the root pom if it has not been
explcitily included.
* Adjust order of command line processing and require of first pom
to ensure command line arguments take precedence. Also, foam.pom POM()
does not override command line set arguments. This was the original
behaviour but was broken in the recent refactor.