DAO Compaction - refactor medusa compaction for single dao compaction
Compaction dumps an MDAO to a new journal, writing each record in
entirety, replacing the original journal which may have a long history
of partial entries as a result of object updates.
Journal replay parsing time is relative to lines parsed, the less lines,
the less time replay takes.
The process supports faceted Sinks to control the compaction of
individual models.